Clarify what "this" means
